Analyzing Jeopardy Questions
In this project, we analyze terminology used in Jeopardy Questions and Answers to see if we can find any patterns that may be helpful to gain an advantage.  We also look into the categories and see if we can narrow down the categories to focus on when studying to gain the highest points.